Refactor of jsonld code into messages
